Azdo provider refactor (#45)
* Refactor to introduce new azdo package

* refactor azdo code to multiple files in azdo package

* Added azdo connection test and refactored provider

* Added azdo provider initial tests

* added more azdo provider tests

* Adding test to verify save env config in azdo provider

* Updating displayed urls to aka.ms
