+++ cli/azd/pkg/azdo/azdo_test.go
@@ -1,0 +1,26 @@
+package azdo
+
+import (
+	"context"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func Test_getAzdoConnection(t *testing.T) {
+	ctx := context.Background()
+	t.Run("empty organization name error", func(t *testing.T) {
+		_, err := GetAzdoConnection(ctx, "", "")
+		assert.EqualError(t, err, "organization name is required")
+	})
+
+	t.Run("empty pat error", func(t *testing.T) {
+		_, err := GetAzdoConnection(ctx, "fake_org", "")
+		assert.EqualError(t, err, "personal access token is required")
+	})
+	t.Run("returns a connection", func(t *testing.T) {
+		connection, err := GetAzdoConnection(ctx, "fake_org", "fake_pat")
+		assert.Nil(t, err)
+		assert.NotNil(t, connection)
+	})
+}
